fix: isolate document review export utilities
